NN-2246: Merge get and post routes into one controller
