+++ backend/controllers/changeCaseload.js
@@ -35,7 +35,26 @@ const changeCaseloadFactory = (elite2Api, logError) => {
     }
   }
 
-  return { index }
+  const post = async (req, res) => {
+    const { caseLoadId } = req.body
+
+    // Don't call API if data is missing
+    if (caseLoadId) {
+      await elite2Api.setActiveCaseload(res.locals, { caseLoadId })
+
+      if (req.session && req.session.userDetails) req.session.userDetails.activeCaseLoadId = caseLoadId
+      if (req.session && req.session.data) req.session.data = null
+
+      res.redirect(config.app.notmEndpointUrl)
+    } else {
+      logError(req.originalUrl, 'Caseload ID is missing')
+      res.render('error.njk', {
+        url: '/change-caseload',
+      })
+    }
+  }
+
+  return { index, post }
 }
 
 module.exports = {
