+++ backend/routes.js
@@ -6,7 +6,6 @@ const { logError } = require('./logError')
 const config = require('./config')
 
 const userCaseLoadsFactory = require('./controllers/usercaseloads').userCaseloadsFactory
-const setActiveCaseLoadFactory = require('./controllers/setactivecaseload').activeCaseloadFactory
 const adjudicationHistoryFactory = require('./controllers/adjudicationHistoryService')
 const offenderServiceFactory = require('./controllers/offenderService')
 const { offenderActivitesFactory } = require('./controllers/offenderActivities')
@@ -86,7 +85,6 @@ const setup = ({ elite2Api, whereaboutsApi, oauthApi, communityApi }) => {
   router.use('/api/me', userMeFactory(oauthApi).userMe)
   router.use('/api/usercaseloads', userCaseLoadsFactory(elite2Api).userCaseloads)
   router.use('/api/userLocations', userLocationsFactory(elite2Api).userLocations)
-  router.post('/api/setactivecaseload', handleErrors(setActiveCaseLoadFactory(elite2Api, logError).setActiveCaseload))
   router.use('/api/houseblockLocations', houseblockLocationsFactory(elite2Api).getHouseblockLocations)
   router.use('/api/activityLocations', activityLocationsFactory(elite2Api).getActivityLocations)
   router.use('/api/bookings/:offenderNo/iepSummary', controller.getIepDetails)
