+++ backend/tests/changeCaseload.test.js
@@ -110,3 +110,70 @@ describe('index', () => {
     })
   })
 })
+
+describe('post', () => {
+  const elite2Api = {}
+  const req = {
+    session: {
+      userDetails: {},
+      data: {
+        activeCaseLoadId: 'LEI',
+      },
+    },
+  }
+  const res = {
+    redirect: jest.fn(),
+    render: jest.fn(),
+  }
+
+  const logError = jest.fn()
+
+  let service
+
+  beforeEach(() => {
+    elite2Api.setActiveCaseload = jest.fn()
+  })
+
+  it('should call the API with the correct body', async () => {
+    service = changeCaseloadFactory(elite2Api, logError)
+
+    req.body = { caseLoadId: 'MDI' }
+
+    await service.post(req, res)
+
+    expect(elite2Api.setActiveCaseload).toHaveBeenCalledWith(res.locals, { caseLoadId: 'MDI' })
+  })
+
+  it('should redirect on success', async () => {
+    service = changeCaseloadFactory(elite2Api, logError)
+
+    req.body = { caseLoadId: 'MDI' }
+
+    await service.post(req, res)
+
+    expect(res.redirect).toHaveBeenCalledWith('//newNomisEndPointUrl/')
+  })
+
+  it('should update session with new caseload id and clear app session data', async () => {
+    service = changeCaseloadFactory(elite2Api, logError)
+
+    req.body = { caseLoadId: 'MDI' }
+
+    await service.post(req, res)
+
+    expect(req.session.data).toBe(null)
+    expect(req.session.userDetails.activeCaseLoadId).toBe('MDI')
+  })
+
+  it('should error when caseload id is missing', async () => {
+    service = changeCaseloadFactory(elite2Api, logError)
+
+    req.body = {}
+
+    await service.post(req, res)
+
+    expect(res.render).toBeCalledWith('error.njk', {
+      url: '/change-caseload',
+    })
+  })
+})
